fix: crash when exiting from PurchaseBaseFragment and no image was taken

+++ app/src/main/java/ch/giantific/qwittig/ui/PurchaseBaseFragment.java
@@ -1301,7 +1301,7 @@ public abstract class PurchaseBaseFragment extends BaseFragment implements
                     }
                 }
             }
-        } else {
+        } else if (!TextUtils.isEmpty(mReceiptImagePath)) {
             boolean fileDeleted = new File(mReceiptImagePath).delete();
             if (!fileDeleted && BuildConfig.DEBUG) {
                 Log.e(LOG_TAG, "failed to delete file");
